Refactored file menu into Game class
